Rename til RisikoscenarioSannsynlighetsBegrunnelseField
